refactor(next): déplace la lecture du cookie darkMode dans Providers (server async) qui délègue à ClientProviders, le RootLayout n'a plus à connaître le mécanisme de persistance
