+++ site/app/layout.tsx
@@ -1,10 +1,5 @@
-import { cookies } from 'next/headers'
 import { ReactNode } from 'react'
 
-import {
-	DARK_MODE_STORAGE_KEY,
-	parseDarkModeValue,
-} from '@/components/utils/darkModeStorage'
 import StyledComponentsRegistry from '@/design-system/StyledComponentsRegistry'
 import { NextJsNavigationProvider } from '@/lib/navigation/providers/NextJsNavigationProvider'
 import i18next, { langue } from '@/locales/i18n-server'
@@ -19,22 +14,13 @@ export const metadata = {
 	),
 }
 
-export default async function RootLayout({
-	children,
-}: {
-	children: ReactNode
-}) {
-	const cookieStore = await cookies()
-	const initialDarkMode = parseDarkModeValue(
-		cookieStore.get(DARK_MODE_STORAGE_KEY)?.value
-	)
-
+export default function RootLayout({ children }: { children: ReactNode }) {
 	return (
 		<html lang={langue}>
 			<body>
 				<StyledComponentsRegistry>
 					<NextJsNavigationProvider>
-						<Providers initialDarkMode={initialDarkMode}>{children}</Providers>
+						<Providers>{children}</Providers>
 					</NextJsNavigationProvider>
 				</StyledComponentsRegistry>
 			</body>
